requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pyyaml>=6.0
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
scipy>=1.11
joblib>=1.3
librosa>=0.10
soundfile>=0.12
pytest>=8.0
